Modernize gemspec file
I generated a new gem and then compared the gemspec there with the one
here and just matched things as it made sense to me. I think there's not
much controversial here.

Note that the docs encourage one to not set `rubygems_version`:

> Do not set this, it is set automatically when the gem is packaged.

https://guides.rubygems.org/specification-reference/#rubygems_version

I also removed some keys that weren't in the docs - I assume they are
stale at this point.